Fix doc source links
At the bottom of each doc page there is a `The source code for this page
can be found here.` link. Currently those use the Paradox default of the
`master` branch, but maligned uses the `main` branch. This change
prevents GitHub from giving a warning about the branch not being found.